Add pytest + CircleCI parallelism (#55)
CircleCI has three default ways that you can split `pytest` tests for parallelization: by file size (literal file size), file/class name (literally splitting by file/class name), or by historical timings for running the tests.

The ultimate goal is to split by historical timings. In order to do that, we first need to have historical data. Thus, to start, we'll split by file size.

Another fun quirk is that the CircleCI executor will fail if the `parallelism` count is higher than the number of "splits". For example, if there is a parallelization factor of 10 but only 3 files with tests, it will fail. For that reason, we have to keep the parallelization count lower than the number of splits. Similarly, it seems to only split by file rather than individual tests; this is an issue because we have many multiple-minute tests per file.

This isn't the best split or parallelization, but it's a step in the right direction.

Adds:
- Parallelization for CircleCI / `pytest`, with an initial parallelization of 2.

Modifies:
- The file structure for integration tests. Refactors the large `test_tools.py` file to multiple files, and adds a `pytest` `conftest.py`.
