+++ toolchest_client/tools/tests/test_sanity.py
@@ -1,6 +1,2 @@
-import pytest
-
-
-@pytest.mark.integration
 def test_sanity():
     assert 1 + 1 == 2
